+++ src/test/scala/firrtlTests/DriverSpec.scala
@@ -8,6 +8,7 @@ import org.scalatest.{FreeSpec, Matchers}
 import firrtl.passes.InlineInstances
 import firrtl.passes.memlib.{InferReadWrite, ReplSeqMem}
 import firrtl._
+import util.BackendCompilationUtilities
 
 class DriverSpec extends FreeSpec with Matchers with BackendCompilationUtilities {
   "CommonOptions are some simple options available across the chisel3 ecosystem" - {
